+++ contracts/TestRecordsMerkleTree.sol
@@ -6,7 +6,7 @@ import "./Rescue.sol";
 import "./RecordsMerkleTree.sol";
 
 contract TestRecordsMerkleTree is RecordsMerkleTree {
-    constructor(uint64 _height) public RecordsMerkleTree(_height) {}
+    constructor(uint8 _height) public RecordsMerkleTree(_height) {}
 
     function testUpdateRecordsMerkleTree(
         uint256[] memory _frontier,
